requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
amqp==2.2.1               # via kombu
billiard==3.5.0.3         # via celery
celery==4.0.2
coverage==4.4.1           # via pytest-cov
ddt==1.1.1
django-celery-results==1.0.1
django-model-utils==3.0.0
freezegun==0.3.9
future==0.16.0
jsonfield==2.0.2
kombu==4.1.0              # via celery
py==1.4.34                # via pytest, pytest-catchlog
pytest-catchlog==1.2.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.1.3             # via pytest-catchlog, pytest-cov, pytest-django
python-dateutil==2.6.1    # via freezegun
python-memcached==1.58
pytz==2017.2              # via celery, django
six==1.10.0               # via freezegun, python-dateutil, python-memcached
vine==1.1.4               # via amqp
mock
